artifacttemplates/https%3A%2F%2Fust-quantil.github.io%2Fartifacttemplates/Qiskit-App-Adapter_DA-w1/files/requirements.txt
Flask==1.1.1
Flask-API==2.0
python-dotenv==0.12.0
